"fixing the build with refactor"
Deleting variantions of goldilocks that we are not using. Adding tests. Cleaniing up and commenting the code
